+++ wabi/base/vt/arrayPyBuffer.cpp
@@ -375,7 +375,7 @@ namespace
       view->format = NULL;
     }
     if ((flags & PyBUF_ND) == PyBUF_ND) {
-      view->ndim = wrapper->shape.size();
+      view->ndim = (int)wrapper->shape.size();
       view->shape = wrapper->shape.data();
     } else {
       view->ndim = 0;
